% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_energyindex_country.R
\name{calculate_energyindex_country}
\alias{calculate_energyindex_country}
\title{Function to obtain the SDEI for selected countries}
\usage{
calculate_energyindex_country(data, method = "fitdis", scale, nvars)
}
\arguments{
\item{data}{dataframe that contains data for one or more countries}

\item{method}{method to calculate the index}

\item{scale}{indicates time scale of aggregation for the index}

\item{nvars}{variable name, e.g., wind production.}
}
\description{
Function to obtain the SDEI for selected countries
}
\details{
whether the index will be calculated on aggregated amounts or not (e.g., daily, 6h)
}
