% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dist.R
\name{choose_dist}
\alias{choose_dist}
\title{Function to estimate the distribution function}
\usage{
choose_dist(xx, mtype, th)
}
\arguments{
\item{xx}{the variable to analysis}

\item{mtype}{predefined distribution}

\item{th}{whether a threshold is needed}
}
\description{
Function to estimate the distribution function
}
