% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ic2017}
\alias{data_ic2017}
\title{Data Definitions for \code{data_ic2017}}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(data_ic2017)
}
\description{
A data frame containing installed wind and solar power capacities for the 27 countries included in data_supply.
}
\examples{
data(data_ic2017)
}
\keyword{datasets}
