% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_energy_drought.R
\name{def_energy_drought}
\alias{def_energy_drought}
\title{Calculation of main features of droughts within the context of renewable energy power systems.}
\usage{
def_energy_drought(dd, mvar, threshval, higher = T, lag = T)
}
\arguments{
\item{dd}{dataframe that contains}

\item{mvar}{variable used to characterise the drought (e.g., PWS or the name of the variable used containing the index values)}

\item{threshval}{threshold values to use when defining droughts}

\item{higher}{is TRUE if the value is above the threshold, and FALSE otherwise}

\item{lag}{to account for the change of sign when examining the occurrence of the drought}
}
\description{
Given a time series of the Standardized energy index obtained with \code{funSDEI},
the function \code{def_energy_drought} returns drought characteristics, including:
}
\references{
Sam Allen and Noelia Otero. 2022. Standardised indices to monitor energy droughts.
}
