% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDis.R
\name{fitDis}
\alias{fitDis}
\title{Function to estimate the distribution function}
\usage{
fitDis(xx, mtype, th)
}
\arguments{
\item{xx}{the variable to analysis}

\item{mtype}{predefined distribution}

\item{th}{whether a threshold is needed}
}
\description{
Function to estimate the distribution function
}
