% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSDEI.R
\name{funSDEI}
\alias{funSDEI}
\title{Function to calculate the index}
\usage{
funSDEI(x, method = c("fitDis", "empirical", "none"), scale)
}
\arguments{
\item{x}{the variable to be standarised}

\item{method}{estimation method}

\item{scale}{indicates whether the variable is aggregate over time}
}
\description{
Function to calculate the index
}
