% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abbrevs.R
\name{get_abbrevs}
\alias{get_abbrevs}
\title{Function to get abbreviations of country names}
\usage{
get_abbrevs(countries)
}
\arguments{
\item{countries}{string with the names of the countries}
}
\description{
Function to get abbreviations of country names
}
