% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drought.R
\name{get_drought}
\alias{get_drought}
\title{Get drought characteristics}
\usage{
get_drought(x, thresholds = c(1.28, 1.64, 1.96), higher = TRUE, lag = FALSE)
}
\arguments{
\item{x}{xts object to be plotted.}

\item{thresholds}{threshold values to use when defining droughts}

\item{higher}{is TRUE if the value is above the threshold, and FALSE otherwise}

\item{lag}{to account for the change of sign when examining the occurrence of the drought}
}
\value{
An xts object containing the drought characteristics corresponding to the input xts.
}
\description{
Extract characteristics of droughts from a time series of standardised index values, including:
}
\details{
Details of the get_drought function will be added here.
}
\examples{
data(data_supply)
# consider hourly German energy supply data in 2019
supply_de <- subset(data_supply, country == "Germany", select = c("date", "PWS"))
supply_de <- xts::xts(supply_de$PWS, order.by = supply_de$date)
options(xts_check_TZ = FALSE)
supply_de_std <- std_index(supply_de, timescale = "hours")

get_drought(supply_de_std, thresholds = c(-1, -1.5, -2), higher = FALSE)

}
\references{
Allen, S. and N. Otero (2023):
`Standardised indices to monitor energy droughts',
\emph{Renewable Energy}
\doi{10.1016/j.renene.2023.119206}

McKee, T. B., Doesken, N. J., & Kleist, J. (1993):
`The relationship of drought frequency and duration to time scales',
\emph{In Proceedings of the 8th Conference on Applied Climatology} 17, 179-183.
}
\author{
Sam Allen, Noelia Otero
}
