% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pit.R
\name{get_pit}
\alias{get_pit}
\title{Calculate probability integral transform values}
\usage{
get_pit(ref_data, new_data, dist = "empirical", return_fit = F)
}
\arguments{
\item{ref_data}{numeric vector from which to estimate the CDF.}

\item{new_data}{numeric vector from which to calculate the PIT values.}

\item{dist}{string; distribution used to estimate the CDF.}

\item{return_fit}{logical; return parameters and goodness-of-fit statistics.}
}
\value{
A vector of PIT values.
}
\description{
Functions to estimate the cumulative distribution function (CDF)
of a set of observations, and return the corresponding probability integral
transform (PIT) values.
}
\details{
Functionality is currently available to calculate parameters using maximum
likelihood estimation, though alternative methods may be introduced later.
}
\examples{
Examples of the calculate_pit function will be added here

}
\author{
Sam Allen, Noelia Otero
}
