% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shortcountry_names.R
\name{get_shortcountry_names}
\alias{get_shortcountry_names}
\title{Function to obtain the ISO names of EU countries}
\usage{
get_shortcountry_names(longname)
}
\arguments{
\item{longname}{one single string with the name of one country}
}
\description{
Function to obtain the ISO names of EU countries
}
