% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ics.R
\name{plot_ics}
\alias{plot_ics}
\title{Function to plot installed capacity values from the data}
\usage{
plot_ics(df, names = NULL)
}
\arguments{
\item{df}{dataframe of the form of data_enerH, with the values of IC wind and solar}

\item{names}{of the countries o be plotted}
}
\description{
Function to plot installed capacity values from the data
}
