% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdei.R
\name{plot_sdei}
\alias{plot_sdei}
\title{Function to plot the index}
\usage{
plot_sdei(dx, yy = seq(1979, 2019), ivar = "SDEI")
}
\arguments{
\item{dx}{dataframe with the SDEI information, dates, index and type of standardized index}

\item{yy}{yeas of study}

\item{ivar}{variable to plot, e.g., index name}
}
\description{
Function to plot the index
}
