% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdei_dist.R
\name{plot_sdei_dist}
\alias{plot_sdei_dist}
\title{Function to plot the index distribution}
\usage{
plot_sdei_dist(
  dx,
  ivar = "SDEI",
  yy = seq(1979, 2019),
  n_bins = 100,
  title = ""
)
}
\arguments{
\item{dx}{dataframe with the SDEI information, dates, index and type of standardized index}

\item{ivar}{index to plot}

\item{yy}{yeas of study}

\item{n_bins}{the number of bins to show}

\item{title}{optional title of the plot}
}
\description{
Function to plot the index distribution
}
