% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDEI_Events.R
\name{SDEI_Events}
\alias{SDEI_Events}
\title{Function to calculate the events based on the SDEI}
\usage{
SDEI_Events(values, type)
}
\arguments{
\item{values}{index value}

\item{type}{?}
}
\description{
Function to calculate the events based on the SDEI
}
