% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_index.R
\name{std_index}
\alias{std_index}
\title{Calculate standardised energy indices}
\usage{
std_index(
  x_new,
  x_ref = x_new,
  timescale = "days",
  dist = "empirical",
  return_fit = FALSE,
  moving_window = NULL,
  window_scale = NULL,
  agg_period = NULL,
  agg_scale = NULL,
  agg_fun = "sum",
  rescale = NULL,
  rescale_fun = "sum",
  index_type = "normal",
  ignore_na = FALSE
)
}
\arguments{
\item{x_new}{vector or time series to be converted to standardised indices.}

\item{x_ref}{vector or time series to be used as reference data when calculating the standardised indices.}

\item{dist}{string; distribution used to calculate the indices.}

\item{return_fit}{logical; return parameters and goodness-of-fit statistics for the distribution fit.}

\item{moving_window}{numeric; length of moving window on which to calculate the indices.}

\item{window_scale}{string; timescale of \code{moving_window}, default is "days".}

\item{agg_period}{numeric; the number of values to aggregate over.}

\item{agg_scale}{string; timescale of \code{agg_period}, default is "days".}

\item{agg_fun}{string; function used to aggregate the data over the aggregation period, default is "sum".}

\item{rescale}{string; the timescale that the time series should be rescaled to.}

\item{rescale_fun}{string; function used to rescale the data, default is "sum".}

\item{index_type}{string; the type of index: "normal" (default), "probability", or "bounded".}

\item{ignore_na}{logical; should NAs be ignored when rescaling the time series?}
}
\value{
Time series of standardised indices.
}
\description{
Inputs a time series of a chosen variable (e.g. residual load or
energy demand) and returns a time series of standardised indices. Different
types of indices can be calculated, on any timescale that is of interest.
}
\details{
Details about the std_index function will be added here
}
\examples{
Examples of the std_index function will be added here

}
\references{
Allen, S. and N. Otero (2022):
`Standardised indices to monitor energy droughts',
\emph{EarthArXiv preprint} 4752.
\doi{10.31223/X51S92}

McKee, T. B., Doesken, N. J., & Kleist, J. (1993):
`The relationship of drought frequency and duration to time scales',
\emph{In Proceedings of the 8th Conference on Applied Climatology} 17, 179-183.
}
\author{
Sam Allen, Noelia Otero
}
