% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_xts.R
\name{aggregate_xts}
\alias{aggregate_xts}
\title{Aggregate values in xts objects}
\usage{
aggregate_xts(
  x,
  agg_length,
  agg_scale = c("days", "hours", "weeks"),
  agg_fun = "sum",
  na_thres = 10
)
}
\arguments{
\item{x}{xts object to be aggregated.}

\item{agg_length}{length of the aggregation period.}

\item{agg_scale}{timescale of the aggregation period, default is 'days'.}

\item{agg_fun}{function to apply to the aggregated data, default is 'sum'.}

\item{na_thres}{threshold for the percentage of NA values allowed in the
aggregation period, default = 10.}
}
\value{
An xts time series with aggregated values.
}
\description{
Inputs an xts time series and outputs an xts time series whose
values have been aggregated over a moving window of a user-specified length.
}
\details{
This has been adapted from code available at
\url{https://github.com/WillemMaetens/standaRdized}.
}
\examples{
Examples of the aggregate_xts function will be added here

}
\author{
Sam Allen, Noelia Otero
}
