% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_supply}
\alias{data_supply}
\title{Data Definitions for \code{data_supply}}
\format{
An object of class \code{"data.frame"}
}
\source{
University of Reading. Dataset.  \doi{10.17864/1947.000321}
}
\usage{
data(data_supply)
}
\description{
This dataset contains wind and solar energy production using the installed capacities
of 2017 considered as a baseline power scenario.
The dataset can be found at https://researchdata.reading.ac.uk/321/, and users are referred to
Bloomfield et al. (2021) for further details.
}
\examples{
data("data_supply")
}
\references{
Bloomfield, Hannah and Brayshaw, David (2021): ERA5 derived time series of European aggregated surface weather variables,
wind power, and solar power capacity factors: hourly data from 1950-2020. \doi{10.17864/1947.000321}
}
\keyword{datasets}
