% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dist.R
\name{fit_dist}
\alias{fit_dist}
\title{Calculate distribution parameters and statistics}
\usage{
fit_dist(data, dist, n_thres = 20)
}
\arguments{
\item{data}{vector of data}

\item{n_thres}{number of data points required to estimate the distribution}

\item{distr}{character string specifying the distribution, see details}
}
\value{
a list containing the estimated distribution function, its parameters,
and relevant goodness-of-fit statistics
}
\description{
This function calculates the distribution parameters and a number
of statistics for a numeric vector and returns them as a named vector.
This function is used internally by the standardized.index function or
can be used to provide input to it.
}
\details{
This has been adapted from code available at
\url{https://github.com/WillemMaetens/standaRdized}.

where relevant, parameter estimation is performed using maximum likelihood
estimation.
}
\examples{
data(Ukkel_RR)

}
