% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSDEI.R
\name{funSDEI}
\alias{funSDEI}
\title{Calculation of Standardised indices to monitor energy droughts}
\usage{
funSDEI(
  x,
  method = c("fitDis", "empirical", "none"),
  scale,
  index_type = c("probability", "bounded", "normal")
)
}
\arguments{
\item{x}{the variable to be standarised}

\item{method}{estimation method}

\item{scale}{indicates whether the variable is aggregate over time}

\item{index_type}{indicates the type of index: "probability", "bounded" or "normal"}
}
\description{
Given a time series of a variable, e.g., wind and solar power generation, returns a standardised indices
based on the method chose. The index can be calculate in any timescale that is of interest (e.g., hourly, daily.)
}
\references{
Sam Allen and Noelia Otero. 2022. Standardised indices to monitor energy droughts.
}
