% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index_UEmap.R
\name{plot_index_UEmap}
\alias{plot_index_UEmap}
\title{Function to plot SEI characteristics in a map}
\usage{
plot_index_UEmap(input_data, fvar)
}
\arguments{
\item{input_data}{dataframe to plot}

\item{fvar}{variable for multiplot, it can be the season}
}
\description{
Function to plot SEI characteristics in a map
}
