% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_index.R
\name{std_index}
\alias{std_index}
\title{Calculate standardised energy indices}
\usage{
std_index(
  x,
  method = c("fitDis", "empirical", "none"),
  scale,
  index_type = c("probability", "bounded", "normal")
)
}
\arguments{
\item{x}{vector or time series to be converted to standardised indices.}

\item{scale}{optional; numeric specifying the number of values to aggregate over.}

\item{index_type}{string; the type of index: "probability", "bounded" or "normal".}

\item{distribution}{string; distribution used to construct the indices.}
}
\value{
Time series of standardised indices.
}
\description{
Inputs a time series of a chosen variable (e.g. residual load or
energy demand) and returns a time series of standardised indices. Different
types of indices can be calculated, on any timescale that is of interest.
}
\details{
Details about the std_index function will be added here
}
\examples{
Examples of the std_index function will be added here

}
\references{
Allen, S. and N. Otero (2022):
`Standardised indices to monitor energy droughts',
\emph{EarthArXiv preprint} 4752.
\doi{10.31223/X51S92}

McKee, T. B., Doesken, N. J., & Kleist, J. (1993):
`The relationship of drought frequency and duration to time scales',
\emph{In Proceedings of the 8th Conference on Applied Climatology} 17, 179-183.
}
\author{
Sam Allen, Noelia Otero
}
